Add estatsd_folsom gen_server
folsom doesn't currently like it when you attempt to add the same
metric concurrently.  The estatsd_folsom server provides
ensure_metric/2 that serializes checks for and, when needed, creation
of metrics.  It does the check by using ets:member on the appropriate
folsom metric ets table which should be less work than folsom's
mechanism of looking through all event handlers.

If this ends up being a bottleneck, one thing to try is to maintain a
separate ets table and enable read concurrency.  Most activity should
be on verifying that metrics keys exist and only rarely creating.