Introduce folsom_blacklist and direct folsom ets communication
The blacklist is specified in app config as a list of regex
(strings).  A key that matches any of the patterns in the blacklist
will not be sent to folsom.

For meter and histogram metrics, instead of going through the event
manager via folsom_metrics:notify, we now call into the appropriate
metric module and do the update directly.  This should save some
messaging overhead.