Reduce race condition window
If a large amount of data were in the ets table, more would be added while the gb_trees object is created. The only risk is that now a crash during flush_metrics would cause all the data to be lost, whereas before it might remain in the ets table (if this gen_server is not the owner because an owner crash deletes the ets table).